backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
behave==1.2.5
behave-django==1.4.0
black==21.12b0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
coverage==6.2
distlib==0.3.4
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
enum34==1.1.10
factory-boy==3.2.1
Faker==11.1.0
filelock==3.4.2
identify==2.4.1
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
nodeenv==1.6.0
parse==1.6.6
parse-type==0.4.0
pathspec==0.9.0
platformdirs==2.4.1
pre-commit==2.16.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.11.2
